Add different error messages for different errors, change all Ints in the GSON files to Longs to better mirror the specification

+++ src/main/kotlin/at/hannibal2/skyhanni/discord/github/GitHubClient.kt
@@ -27,13 +27,13 @@ class GitHubClient(owner: String, repo: String, private val token: String) {
         }
     }
 
-    fun downloadArtifact(artifactId: Int, outputFile: File) {
+    fun downloadArtifact(artifactId: Long, outputFile: File) {
         readBody("$base/actions/artifacts/$artifactId/zip") { body ->
             outputFile.writeBytes(body.bytes())
         }
     }
 
-    fun findPullRequest(prNumber: Int): PullRequestJson? {
+    fun findPullRequest(prNumber: Long): PullRequestJson? {
         return readJson<PullRequestJson, PullRequestJson?>("$base/pulls/$prNumber") { it }
     }
